Moved STTextviewController to new folder. Moved cursor logic to STTextviewController+Cursor
