+++ Sources/CodeEditTextView/Controller/STTextViewController.swift
@@ -46,6 +46,9 @@ public class STTextViewController: NSViewController, STTextViewDelegate, ThemeAt
     /// The font to use in the `textView`
     public var font: NSFont
 
+    /// The current cursor position e.g. (1, 1)
+    public var cursorPosition: Binding<(Int, Int)>
+
     /// The editorOverscroll to use for the textView over scroll
     public var editorOverscroll: Double
 
@@ -328,102 +331,6 @@ public class STTextViewController: NSViewController, STTextViewDelegate, ThemeAt
     override public func keyDown(with event: NSEvent) {
         // TODO: - This should be uncessecary
     }
-
-    // MARK: Cursor Position
-
-    private var cursorPosition: Binding<(Int, Int)>
-
-    private func setCursorPosition(_ position: (Int, Int)) {
-        guard let provider = textView.textLayoutManager.textContentManager else {
-            return
-        }
-
-        var (line, column) = position
-        let string = textView.string
-        if line > 0 {
-            if string.isEmpty {
-                // If the file is blank, automatically place the cursor in the first index.
-                let range = NSRange(string.startIndex..<string.endIndex, in: string)
-                if let newRange = NSTextRange(range, provider: provider) {
-                    _ = self.textView.becomeFirstResponder()
-                    self.textView.setSelectedRange(newRange)
-                    return
-                }
-            }
-
-            string.enumerateSubstrings(in: string.startIndex..<string.endIndex) { _, lineRange, _, done in
-                line -= 1
-                if line < 1 {
-                    // If `column` exceeds the line length, set cursor to the end of the line.
-                    let index = min(lineRange.upperBound, string.index(lineRange.lowerBound, offsetBy: column - 1))
-                    if let newRange = NSTextRange(NSRange(index..<index, in: string), provider: provider) {
-                        self.textView.setSelectedRange(newRange)
-                    }
-                    done = true
-                } else {
-                    done = false
-                }
-            }
-        }
-    }
-
-    private func updateCursorPosition() {
-        guard let textLayoutManager = textView.textLayoutManager as NSTextLayoutManager?,
-              let textContentManager = textLayoutManager.textContentManager as NSTextContentManager?,
-              let insertionPointLocation = textLayoutManager.insertionPointLocation,
-              let documentStartLocation = textLayoutManager.documentRange.location as NSTextLocation?,
-              let documentEndLocation = textLayoutManager.documentRange.endLocation as NSTextLocation?
-        else {
-            return
-        }
-
-        let textElements = textContentManager.textElements(
-            for: NSTextRange(location: textLayoutManager.documentRange.location, end: insertionPointLocation)!)
-        var line = textElements.count
-
-        textLayoutManager.enumerateTextSegments(
-                in: NSTextRange(location: insertionPointLocation),
-                type: .standard,
-                options: [.rangeNotRequired, .upstreamAffinity]
-            ) { _, textSegmentFrame, _, _ -> Bool
-                in
-                var col = 1
-                /// If the cursor is at the end of the document:
-                if textLayoutManager.offset(from: insertionPointLocation, to: documentEndLocation) == 0 {
-                    /// If document is empty:
-                    if textLayoutManager.offset(from: documentStartLocation, to: documentEndLocation) == 0 {
-                        self.cursorPosition.wrappedValue = (1, 1)
-                        return false
-                    }
-                    guard let cursorTextFragment = textLayoutManager.textLayoutFragment(for: textSegmentFrame.origin),
-                          let cursorTextLineFragment = cursorTextFragment.textLineFragments.last
-                    else { return false }
-
-                    col = cursorTextLineFragment.characterRange.length + 1
-                    if col == 1 { line += 1 }
-                } else {
-                    guard let cursorTextLineFragment = textLayoutManager.textLineFragment(at: insertionPointLocation)
-                    else { return false }
-
-                    /// +1, because we start with the first character with 1
-                    let tempCol = cursorTextLineFragment.characterIndex(for: textSegmentFrame.origin)
-                    let result = tempCol.addingReportingOverflow(1)
-
-                    if !result.overflow { col = result.partialValue }
-                    /// If cursor is at end of line add 1:
-                    if cursorTextLineFragment.characterRange.length != 1 &&
-                        (cursorTextLineFragment.typographicBounds.width == (textSegmentFrame.maxX + 5.0)) {
-                        col += 1
-                    }
-
-                    /// If cursor is at first character of line, the current line is not being included
-                    if col == 1 { line += 1 }
-                }
-
-                self.cursorPosition.wrappedValue = (line, col)
-                return false
-            }
-        }
 
     deinit {
         textView = nil
